relocate components and patterns to documented folder structure
