+++ app/frontend/src/components/loader.scss
