+++ app/frontend/src/components/panel.test.js
