+++ app/frontend/src/components/searchCheckbox.test.js
