+++ app/frontend/src/patterns/autocomplete.scss
